% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.Genotype.R
\name{GAPIT.Genotype}
\alias{GAPIT.Genotype}
\title{GAPIT.Genotype}
\usage{
GAPIT.Genotype(
  G = NULL,
  GD = NULL,
  GM = NULL,
  KI = NULL,
  kinship.algorithm = "Zhang",
  SNP.effect = "Add",
  SNP.impute = "Middle",
  PCA.total = 0,
  PCA.col = NULL,
  PCA.3d = PCA.3d,
  seed = 123,
  SNP.fraction = 1,
  file.path = NULL,
  file.from = NULL,
  file.to = NULL,
  file.total = NULL,
  file.fragment = 1000,
  SNP.test = TRUE,
  file.G = NULL,
  file.Ext.G = NULL,
  file.GD = NULL,
  file.Ext.GD = NULL,
  file.GM = NULL,
  file.Ext.GM = NULL,
  SNP.MAF = 0.05,
  FDR.Rate = 0.05,
  SNP.FDR = 1,
  Timmer = NULL,
  Memory = NULL,
  LD.chromosome = NULL,
  LD.location = NULL,
  LD.range = NULL,
  SNP.CV = NULL,
  GP = NULL,
  GK = NULL,
  GTindex = NULL,
  bin.size = 1000,
  inclosure.size = 100,
  sangwich.top = NULL,
  sangwich.bottom = NULL,
  file.output = TRUE,
  kinship.cluster = "average",
  NJtree.group = NULL,
  NJtree.type = c("fan", "unrooted"),
  Create.indicator = FALSE,
  Major.allele.zero = FALSE,
  Geno.View.output = TRUE
)
}
\arguments{
\item{G}{matrix of genotypes in HapMap format}

\item{GD}{matrix of genotypes in numeric format}

\item{GM}{Genotype Map for numeric data}

\item{KI}{Kinship matrix}

\item{kinship.algorithm}{myDescription}

\item{SNP.effect}{myDescription}

\item{SNP.impute}{myDescription}

\item{PCA.total}{myDescription}

\item{PCA.col}{myDescription}

\item{PCA.3d}{myDescription}

\item{seed}{myDescription}

\item{SNP.fraction}{myDescription}

\item{file.path}{myDescription}

\item{file.from}{myDescription}

\item{file.to}{myDescription}

\item{file.total}{myDescription}

\item{file.fragment}{myDescription}

\item{SNP.test}{myDescription}

\item{file.G}{myDescription}

\item{file.Ext.G}{myDescription}

\item{file.GD}{myDescription}

\item{file.Ext.GD}{myDescription}

\item{file.GM}{myDescription}

\item{file.Ext.GM}{myDescription}

\item{SNP.MAF}{myDescription}

\item{FDR.Rate}{myDescription}

\item{SNP.FDR}{myDescription}

\item{Timmer}{myDescription}

\item{Memory}{myDescription}

\item{LD.chromosome}{myDescription}

\item{LD.location}{myDescription}

\item{LD.range}{myDescription}

\item{SNP.CV}{myDescription}

\item{GP}{myDescription}

\item{GK}{myDescription}

\item{GTindex}{myDescription}

\item{bin.size}{myDescription}

\item{inclosure.size}{myDescription}

\item{sangwich.top}{myDescription}

\item{sangwich.bottom}{myDescription}

\item{file.output}{myDescription}

\item{kinship.cluster}{myDescription}

\item{NJtree.group}{myDescription}

\item{NJtree.type}{myDescription}

\item{Create.indicator}{myDescription}

\item{Major.allele.zero}{myDescription}

\item{Geno.View.output}{myDescription}
}
\value{
a list
}
\description{
Genotype function
}
\details{
More to come.
}
\examples{
# GAPIT.Genotype()

}
\author{
Zhiwu Zeng, Jiabo Wang
}
