% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.Numericalization.R
\name{GAPIT.Numericalization}
\alias{GAPIT.Numericalization}
\title{GAPIT.Numericalization}
\usage{
GAPIT.Numericalization(
  x,
  bit = 2,
  effect = "Add",
  impute = "None",
  Create.indicator = FALSE,
  Major.allele.zero = FALSE,
  byRow = TRUE
)
}
\arguments{
\item{x}{a data.frame containing HapMap formatted data}

\item{bit}{myDescription}

\item{effect}{myDescription}

\item{impute}{myDescription}

\item{Create.indicator}{myDescription}

\item{Major.allele.zero}{myDescription}

\item{byRow}{myDescription}
}
\value{
a list
}
\description{
Convert from HapMap to numerical format
}
\details{
More to come.

HapMap has a header but we set header == FALSE???
}
\examples{
# GAPIT.Numericalization()

}
\author{
Feng Tian and Zhiwu Zhang
}
