% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.R
\name{GAPIT}
\alias{GAPIT}
\title{GAPIT Genome Association and Prediction Integrated Tools}
\usage{
GAPIT(
  Y = NULL,
  G = NULL,
  GD = NULL,
  GM = NULL,
  KI = NULL,
  Z = NULL,
  CV = NULL,
  Aver.Dis = 1000,
  a2 = 0,
  adim = 2,
  acceleration = 0,
  alpha = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  buspred = FALSE,
  bin.from = 10000,
  bin.to = 10000,
  bin.by = 10000,
  bin.size = c(1e+06),
  bin.selection = c(10, 20, 50, 100, 200, 500, 1000),
  BINS = 20,
  converge = 1,
  cutOff = 0.05,
  category = 1,
  cveff = NULL,
  Create.indicator = FALSE,
  CG = NULL,
  CV.Inheritance = NULL,
  DPP = 1e+05,
  DP = NULL,
  esp = 1e-10,
  effectunit = 1,
  file.from = 1,
  file.to = 1,
  file.total = NULL,
  file.fragment = 99999,
  file.path = NULL,
  file.G = NULL,
  file.Ext.G = NULL,
  file.GD = NULL,
  file.GM = NULL,
  file.Ext.GD = NULL,
  file.Ext.GM = NULL,
  file.output = TRUE,
  FDR.Rate = 1,
  FDRcut = FALSE,
  group.from = 1e+06,
  group.to = 1e+06,
  group.by = 20,
  GTindex = NULL,
  Geno.View.output = TRUE,
  GP = NULL,
  GK = NULL,
  h2 = NULL,
  inclosure.from = 10,
  inclosure.to = 10,
  inclosure.by = 10,
  iteration.output = FALSE,
  iteration.method = "accum",
  inpch = NULL,
  Inter.Plot = FALSE,
  Inter.type = c("m", "q"),
  kinship.cluster = "average",
  kinship.group = "Mean",
  kinship.algorithm = "Zhang",
  llim = -10,
  lmpred = FALSE,
  LD.chromosome = NULL,
  LD.location = NULL,
  LD.range = NULL,
  LD = 0.1,
  model = "MLM",
  method.GLM = "FarmCPU.LM",
  method.sub = "reward",
  method.sub.final = "reward",
  method.bin = "static",
  maxOut = 100,
  memo = NULL,
  maxLoop = 3,
  Model.selection = FALSE,
  Multi_iter = FALSE,
  Major.allele.zero = FALSE,
  Multiple_analysis = TRUE,
  num_regwas = 10,
  ncpus = 1,
  ngrid = 100,
  NQTN = NULL,
  N.sig = NULL,
  NJtree.group = NULL,
  NJtree.type = c("fan", "unrooted"),
  opt = "extBIC",
  output.numerical = FALSE,
  output.hapmap = FALSE,
  outpch = NULL,
  QTN = NULL,
  QTN.round = 1,
  QTN.limit = 0,
  QTN.update = TRUE,
  QTN.method = "Penalty",
  QC = TRUE,
  QC.Y = FALSE,
  QTN.position = NULL,
  QTNDist = "normal",
  r = 0.25,
  Random.model = TRUE,
  sangwich.top = NULL,
  sangwich.bottom = NULL,
  seed = NULL,
  SNP.P3D = TRUE,
  SNP.effect = "Add",
  SNP.impute = "Middle",
  SNP.fraction = 1,
  SNP.test = TRUE,
  SNP.MAF = 0,
  SNP.FDR = 1,
  SNP.permutation = FALSE,
  SNP.CV = NULL,
  SNP.robust = "GLM",
  SUPER_GD = NULL,
  SUPER_GS = FALSE,
  threshold.output = 0.01,
  testY = NULL,
  plot.style = "Oceanic",
  plot.bin = 10^5,
  PCA.total = 0,
  PCA.col = NULL,
  PCA.3d = FALSE,
  PCA.legend = NULL,
  PCA.View.output = TRUE,
  Phenotype.View = TRUE,
  Prior = NULL,
  Para = NULL,
  ulim = 10,
  WS = c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),
  WS0 = NULL,
  ws = 20
)
}
\arguments{
\item{Y}{= NULL, data.frame of phenotype data, samples in rows, traits in column, first column is sample name}

\item{G}{= NULL, data.frame of genotypic data, HAPMAP format}

\item{GD}{= NULL, data.frame of genetic data in 'numerical' format, samples in rows, variants in columns.}

\item{GM}{= NULL, Genetic Map data.frame to provide genomic coordinates for GD}

\item{KI}{= NULL, Kinship matrix}

\item{Z}{= NULL,}

\item{CV}{= NULL, Covariate matrix}

\item{a2}{= 0,}

\item{adim}{= 2,}

\item{acceleration}{= 0,}

\item{alpha}{= c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),}

\item{buspred}{= FALSE,}

\item{bin.from}{= 10000,}

\item{bin.to}{= 10000,}

\item{bin.by}{= 10000,}

\item{bin.size}{= c(1e+06),}

\item{bin.selection}{= c(10, 20, 50, 100, 200, 500, 1000),}

\item{BINS}{= 20,}

\item{converge}{= 1,}

\item{cutOff}{= 0.05,}

\item{category}{= 1,}

\item{cveff}{= NULL,}

\item{Create.indicator}{= FALSE,}

\item{CG}{= NULL,}

\item{CV.Inheritance}{= NULL,}

\item{DPP}{= 1e+05,}

\item{esp}{= 1e-10,}

\item{effectunit}{= 1,}

\item{file.from}{= 1,}

\item{file.to}{= 1,}

\item{file.total}{= NULL,}

\item{file.fragment}{= 99999,}

\item{file.path}{= NULL,}

\item{file.G}{= NULL,}

\item{file.Ext.G}{= NULL,}

\item{file.GD}{= NULL,}

\item{file.GM}{= NULL,}

\item{file.Ext.GD}{= NULL,}

\item{file.Ext.GM}{= NULL,}

\item{file.output}{= TRUE,}

\item{FDR.Rate}{= 1,}

\item{FDRcut}{= FALSE,}

\item{group.from}{= 1e+06,}

\item{group.to}{= 1e+06,}

\item{group.by}{= 20,}

\item{GTindex}{= NULL,}

\item{Geno.View.output}{= TRUE,}

\item{GP}{= NULL,}

\item{GK}{= NULL,}

\item{h2}{= NULL,}

\item{inclosure.from}{= 10,}

\item{inclosure.to}{= 10,}

\item{inclosure.by}{= 10,}

\item{iteration.output}{= FALSE,}

\item{iteration.method}{= "accum",}

\item{Inter.Plot}{= FALSE,}

\item{Inter.type}{= c("m", "q"),}

\item{kinship.cluster}{= "average", options: complete, ward, single, mcquitty, median, and centroid}

\item{kinship.group}{= "Mean", options: Max, Min, and Median}

\item{kinship.algorithm}{= "VanRaden", options: EMMA, Loiselle, VanRaden, Zhang}

\item{llim}{= -10,}

\item{lmpred}{= FALSE,}

\item{LD.chromosome}{= NULL,}

\item{LD.location}{= NULL,}

\item{LD.range}{= NULL,}

\item{LD}{= 0.1,}

\item{model}{= "MLM", options: MLM, GLM, CMLM, MMLM, SUPER, FarmCPU, gBLUP, or cBLUP}

\item{method.GLM}{= "FarmCPU.LM",}

\item{method.sub}{= "reward",}

\item{method.sub.final}{= "reward",}

\item{method.bin}{= "static",}

\item{maxOut}{= 100,}

\item{memo}{= NULL,}

\item{maxLoop}{= 3,}

\item{Model.selection}{= FALSE,}

\item{Multi_iter}{= FALSE,}

\item{Major.allele.zero}{= FALSE,}

\item{Multiple_analysis}{= FALSE,}

\item{num_regwas}{= 10,}

\item{ncpus}{= 1,}

\item{ngrid}{= 100,}

\item{NQTN}{= NULL,}

\item{NJtree.group}{= NULL,}

\item{NJtree.type}{= c("fan", "unrooted"),}

\item{opt}{= "extBIC",}

\item{output.numerical}{= FALSE,}

\item{output.hapmap}{= FALSE,}

\item{QTN}{= NULL,}

\item{QTN.round}{= 1,}

\item{QTN.limit}{= 0,}

\item{QTN.update}{= TRUE,}

\item{QTN.method}{= "Penalty",}

\item{QC}{= TRUE,}

\item{QTN.position}{= NULL,}

\item{QTNDist}{= "normal",}

\item{r}{= 0.25,}

\item{Random.model}{= FALSE,}

\item{sangwich.top}{= NULL,}

\item{sangwich.bottom}{= NULL,}

\item{seed}{= NULL,}

\item{SNP.P3D}{= TRUE,}

\item{SNP.effect}{= "Add",}

\item{SNP.impute}{= "Middle",}

\item{SNP.fraction}{= 1,}

\item{SNP.test}{= TRUE,}

\item{SNP.MAF}{= 0,}

\item{SNP.FDR}{= 1,}

\item{SNP.permutation}{= FALSE,}

\item{SNP.CV}{= NULL,}

\item{SNP.robust}{= "GLM",}

\item{SUPER_GD}{= NULL,}

\item{SUPER_GS}{= FALSE,}

\item{threshold.output}{= 0.01,}

\item{testY}{= NULL,}

\item{plot.style}{= "Oceanic",}

\item{plot.bin}{= 10^5,}

\item{PCA.total}{= 0,}

\item{PCA.col}{= NULL,}

\item{PCA.3d}{= FALSE,}

\item{PCA.View.output}{= TRUE,}

\item{Prior}{= NULL,}

\item{Para}{= NULL}

\item{ulim}{= 10,}

\item{WS}{= c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),}

\item{WS0}{= 1e6,}
}
\value{
A list
including some of the following elements:MLM, GLM, CMLM, MMLM, SUPER, FarmCPU, gBLUP, cBLUP
}
\description{
GAPIT analyzes phenotypic and genotypics data to infer association.
}
\details{
Genome Association and Prediction Integrated Tools
Available models: MLM, GLM, CMLM, MMLM, SUPER, FarmCPU, gBLUP, cBLUP
}
\examples{
\dontrun{

myPhenoFile <- system.file("extdata", "mdp_traits.txt.gz", package = "GAPIT3")
myGenoFile <- system.file("extdata", "mdp_genotype_test.hmp.txt.gz", package = "GAPIT3")
myPhenotypes <- read.table(myPhenoFile, header = TRUE)
myGenotypes  <- read.table(myGenoFile, header = FALSE)

myGAPIT <- GAPIT(
  Y = myPhenotypes,
  G = myGenotypes,
  PCA.total = 3,
  file.output = FALSE,
  model = "MLM"
)
}


}
\seealso{
GAPIT.DP(), GAPIT.Phenotype.View(), GAPIT.judge(), GAPIT.IC(), GAPIT.SS(), GAPIT.ID().


library(help = "GAPIT3")
}
\author{
Zhiwu Zhang and Jiabo Wang
}
